% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2pdata.r
\name{as.2pdata}
\alias{as.2pdata}
\title{Converts a data frame to a list made for fitting LBH model}
\usage{
as.2pdata(f_pos, f_zero = f_pos, f_area, data)
}
\arguments{
\item{f_pos}{an object of class \code{\link[stats]{formula}}:
a symbolic description of the fixed effect model to be fitted to the positive part.}

\item{f_zero}{an object of class \code{\link[stats]{formula}}:
a symbolic description of the fixed effect model to be fitted to the binary part.
Default value is to using the same formula as the positive part (\code{f_pos}).}

\item{f_area}{an object of class \code{\link[stats]{formula}}:
a symbolic description of the area code to be fitted to both the positive part and the negative part.}

\item{data}{data frame containing the variables named in \code{f_pos}, \code{f_zero} and \code{f_area}.}
}
\value{
An object of the class "2pdata" which is a list with the following components:
\itemize{
  \item \code{lys}: response vector in log scale for the positive part
  \item \code{Xs1}: model matrix for the positive part
  \item \code{deltas}: binary response vector for the binary part
  \item \code{Xs0}: model matrix for the binary part
  \item \code{area}: vector with the area code
}
}
\description{
The default method deals with the \code{NA} values in the input data and
transforms the data into a format that fits the framework of the unit level model of Lyu, Berg and Hofmann.
}
\details{
The response variable in the formula \code{f_zero} is ignored.
  \code{I(y>0)} will be used for the binary part where \code{y} is the
  response variable in the formula \code{f_pos}.
}
\examples{
  erosion_2p <- as.2pdata(f_pos = RUSLE2~logR+logK+logS,
                          f_zero = ~logR+logS+crop2+crop3,
                          f_area = ~cty, data = erosion)
}
