% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{erosion}
\alias{erosion}
\title{Fake erosion data}
\format{A data frame with 646 rows and 10 variables.
\itemize{
  \item{ctylab}: {county name}
  \item{cty}: {county code}
  \item{mukey}: {soil map unit key}
  \item{crop}: {crop category from the CDL data}
  \item{logR}: {\code{log} USLE rainfall factor}
  \item{logK}: {\code{log} USLE erosion erodibility factor}
  \item{logS}: {\code{log} USLE soil slope factor}
  \item{crop2}: {indicator of soybean}
  \item{crop3}: {indicator of spring wheat}
  \item{RUSLE2}: {soil erosion loss in tons per year}
}}
\usage{
erosion
}
\description{
Fake data on soil erosion and other related variables for 64 South Dakota counties.
}
\seealso{
\code{\link{Xaux}}
}
\keyword{datasets}
