% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{map_sd}
\alias{map_sd}
\title{South Dakota map data}
\format{An object of class \code{data.frame} with 1414 rows and 8 columns.}
\usage{
map_sd
}
\description{
South Dakota county map coordinates in longitude and latitude with FIPS.
}
\keyword{datasets}
