% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebLBH.r
\name{ebLBH}
\alias{ebLBH}
\title{EB estimator of area means and associated MSE estimator}
\usage{
ebLBH(Xnonsample, f_q = ~1, data_2p, fit)
}
\arguments{
\item{Xnonsample}{matrix or data frame containing covariates, the area code and
the variables named in \code{f_q} for the out-of-sample units.}

\item{f_q}{an object of class \code{\link[stats]{formula}}:
a symbolic description of the number of out-of-sample units with the same covariates.
Default value is \code{~1}.}

\item{data_2p}{a two-part data object returned by \code{\link{as.2pdata}}.}

\item{fit}{a list from model fitting with components as returned by \code{\link{mleLBH}}.}
}
\value{
A data frame with the number of rows equal to the number of unique areas in \code{Xnonsample}:
\itemize{
 \item \code{area}: area codes
 \item \code{eb}: EB estimator of area means
 \item \code{mse}: one-step MSE estimator of area means
}
}
\description{
Obtains numerical approximations of EB estimators of area means
under the unit level model of Lyu, Berg and Hofmann
when the values of auxiliary variables for out-of-sample units and
the model parameter estimates are available.
}
\examples{
  erosion_2p <- as.2pdata(f_pos = RUSLE2~logR+logK+logS,
                          f_zero = ~logR+logS+crop2+crop3,
                          f_area = ~cty, data = erosion)
  fit <- mleLBH(erosion_2p)
  predictions <- ebLBH(Xaux, f_q = ~cnt, erosion_2p, fit)
}
