% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simLBH.r
\name{simLBH}
\alias{simLBH}
\title{Simulate unit responses under the unit level model of Lyu, Berg and Hofmann}
\usage{
simLBH(fit, Xs, f_pos, f_zero = f_pos, f_area, link = "logit")
}
\description{
Simulate responses given the model parameter and covariates.
}
\examples{
  erosion_2p <- as.2pdata(f_pos = RUSLE2~logR+logK+logS,
                          f_zero = ~logR+logS+crop2+crop3,
                          f_area = ~cty, data = erosion)
  fit <- mleLBH(erosion_2p)
  responses <- simLBH(fit, Xaux, f_pos = RUSLE2~logR+logK+logS,
                      f_zero = ~logR+logS+crop2+crop3, f_area = ~cty)
}
